+++ AopCaching/CacheListAttribute.cs
@@ -269,7 +269,7 @@ namespace PubComp.Caching.AopCaching
         {
             if (Interlocked.Read(ref initialized) == 0L)
             {
-                this.cache = await CacheManager.GetCacheAsync(this.cacheName).ConfigureAwait(false);
+                this.cache = CacheManager.GetCache(this.cacheName);
                 Interlocked.Exchange(ref initialized, 1L);
             }